Kill PV.parseVersionConstraint
It parses “=1.0.0” but not “=1.0.0 || 2.0.0”.  It shouldn’t be a public thing (and is never used publicly).

There should be only two things: versions and constraints.  A value that can legally be “=1.0.0” is not a version, it is a constraint (and yet we still call it versionString in a few places).  A value that can be “=1.0.0” but not “1.0.0 || 2.0.0” is not a constraint; it is at best a rare creature called something like a “simple constraint.”  We may expand the syntax of constraints again in the future, and we can’t have N things called a “constraint.”