use Commonmark to process READMEs
Use the Commonmark NPM package to process README files.

Commonmark is imperfect -- it has at least one bug that we work around in code.
It is probably still useful to do this for the longterm, since it will make it
easier to do more detailed processing iff we ever decide to.