always change metadata
Don't play clever tricks with blank/missing options -- always change
the metadata on disk to the contents of the package.js file. I think that's
actually a better user experience, because it is simpler (even if less
customizable).

Also make sure that uploadFile returns an error code if it fails, and use
that to exit the buildmessage when needed.