Add ––spy option to self-test (WIP)
This option should log activity as it happens, including commands run
and their output, so you can see what is happening (and what is taking
so long).

Not finished, because there are more commands to add logging to
(like “mkdir”, “unlink”, “set”, …).  Also, the printing doesn’t seem
to interleave with the status spinner correctly.