Create CS.Input and clean up resolver options
CS.Input is a serializable representation of the “problem.”  It includes
the arguments to PackagesResolver#resolve, and also the catalog data
loaded into the CatalogCache.  It’s independent of the solver, and
doesn’t even know about PackagesResolver or Resolver.

Along the way, get rid of the _testing and _debug flags.  “_testing”
came about to avoid running the real cost function on some of the unit
tests, but it doesn’t actually seem to matter anymore for correctness
or performance of the tests.  “_debug” was just used to enable some
console.logs, and possibly shouldn’t have been committed in the first
place.