Reproduce slow version solving in a test
This unit test demonstrates 20-second solving time.  Thanks to the
CatalogCache abstraction, the data provided to the solver in the test
is exactly the data it gets when running the “meteor” command in a test
app with a bunch of packages in .meteor/packages and no .meteor/versions
file.

The test is hidden behind an environment variable:
CONSTRAINT_SOLVER_SLOW_TESTS