Always extract user-readable,listable directories from tarballs
This works around the fact that a directory tarred up on Windows will
untar without an executable bit on Mac/Linux (because there is no such
thing as an executable permission for directories on Windows). So we
explicitly set directories to be readable and executable by the owner
when untarring.

This fix follows the pattern that npm uses, though npm uses 0777 instead
of 0500. That seems overly permissive to me; I'm not sure why we would
need to explicitly set it to writeable too (and readable/listable by
other users).