Lift isSatisfied out of resolver.js
If we axe the rest of resolver.js, we still need this piece.

It would be as simple as VersionConstraint#isSatisfiedBy(version)
except for the fact that we put the logic here about whether to allow
prereleases, which necessitates the “package” and “options” arguments
and makes this a constraint-solver-specific utility.  In the future,
we can put that logic somewhere else, but for now keeping it
preserves the current behavior.