Tests for `Template.currentData` and `Template.parentData`
A recent change fixed the behavior of `Template.currentData`
and `Template.parentData` in event handlers and helpers.
These are tests for the new, correct behavior.

(The old behavior read the data of the template instance.
The new behavior read the data context around the DOM node
where the event or helper ran)