meteor login --email was supposed to be boolean!
See #3532.

We honestly should probably just drop the option and allow you to type
either a username or a password, just like in accounts-ui.

Because options have to be consistently bool or not across commands (so
that you can put them before the command name), change the testing-only
dummy command's string option to a different nonsense name.