Improve cost function
It's still not really correct to do it this way, but it passes the
existing tests now.

In particular, we now prioritize making top-level dependencies
newer over making indirect dependencies older.  However, this
approach is flawed because we only visit the packages in the
original solution, before optimization.  We should assign a cost
to every package version that might be chosen, and let the solver
shift things around arbitrarily.