Support getting MiniSat's conflict clause
It really just says which assumptions created the conflict when
solveAssuming is used.  And currently we only support one assumption
(which may represent a formula).  If we supported N assumptions, it's
possible MiniSat's analysis would tell us which of the N is the
problem.  It's not clear if it would be a minimal set though.

Still, might as well have the capability at the level of interfacing
with MiniSat in case it comes in handy.