Steps towards explaining version solver errors
The scheme of relaxing as few constraints as possible seems to work.

There are only two types of errors: "unknown package" and "conflict".

If we ignore all version constraints and the problem is still
unsatisfiable, it must be because there's a package that's
unavoidably required that doesn't exist, i.e. an "unknown package".

Otherwise, if there are N constraints in the entire problem graph,
we know the problem is satisfiable with 0 of them in force,
and unsatisfiable with all N of them in force.  We use the SAT solver
to find a solution that maximizes the number of constraints in force.
Usually there is one with N-1 constraints in force, and we can
analyze the constraint the breaks the camel's back in order to report
a conflict to the user.

At the moment the code prints to the console about its findings, and
doesn't yet generate a nice error.  Needs code to generate the paths
by which chosen packages are reached, too.  And maybe we should
still run the version optimization so that we don't confuse the
user by reporting conflicts in old versions?  For example, if all
versions of "foo" require a problematic version of "bar", it would be
better to report a problem with "foo@1.0.0" than with "foo@0.0.1".