Refactor server/shell.js into classes, and add simple authentication.
Summary:
Requiring a key that is only discoverable via the file system will allow
us to safely listen for connections on a port instead of a socket file.

Test Plan:
Use `meteor shell` and verify that it still works as before, especially
when started before/after/during server startup, and when the server is
killed while multiple shell clients are connected.

Reviewers: Slava, dgreensp, stubailo, glasser

Reviewed By: glasser

Differential Revision: https://phabricator.meteor.io/D20