Use a port instead of a socket file for `meteor shell`.
Summary:
This eliminates a whole class of problems relating to socket files (such
as the limit on the length of their names), and potentially makes it
easier to implement `meteor shell` on other platforms, such as Windows.

Fixes #3591.

Test Plan:
Use `meteor shell` and verify that it still works as before, especially
when started before/after/during server startup, and when the server is
killed while multiple shell clients are connected.

Reviewers: Slava, dgreensp, stubailo, glasser

Reviewed By: glasser

Differential Revision: https://phabricator.meteor.io/D22