Silently disable shell access if socket filename is too long.
Summary:
Note that an explanation will be printed to the console for clients
attempting to connect via `meteor shell`, but nothing will be printed to
the server console.

Fixes #3591.

Test Plan:
Artificially lower the limit and verify that the server starts as usual,
but that shell clients exit with a warning.

Reviewers: Slava, dgreensp, stubailo, glasser

Reviewed By: glasser

Differential Revision: https://phabricator.meteor.io/D21