Add extra note about enabling internal packages
This simply adds an extra note to the 'writing packages' section of the docs which states that you should `meteor add` an internal package to enable it. I ran into this issue and banged my head for a while last night wondering why an internal package was not exposing a template / exported object to the rest of my app. The issue was that I did not `meteor add` it since I assumed that because it was internal source code I did not have to.