Fix selftests that create packages with colons
We used to create package directories with the same name as the package
name, but on windows you can't have directories with colons in them.

Now Sandbox.prototype.createPackage takes an additional argument
with the directory name, as distinct from the package name.

This commit updates all the tests that called createPackage to generate
a directory name with no colons.