Unit tests of previous solution cost function
After penalizing incompatible versions, we want
to get as close as possible to the previous solution,
with the one wrinkle that we want to resist patches
for compatible versions but not for incompatible versions.
If we actually do break a previous version, we might as
well take patches while we're at it.