Implement new prerelease handling
The new rules are:
- "anticipated prereleases" get a totally free pass, as before
- number of unanticipated prereleases is minimized
- if we end up using any, it sets the neededToUse... flag on return

Removed the nuance where if a constraint mentions any prerelease, then
other prereleases of that package are locally considered "anticipated".
It was basically an escape valve because we couldn't do minimization.
I actually was about to emulating it anyway -- specifically the fact
that if a constraint asks for -pre.0 (which is "anticipated") and the
solver selects -pre.1 (which is not), the solver does NOT set the
neededToUse flag because of the "locally anticipated" exemption.
However, this is not principled anyway because it clashes with the
tool's assessment, which will still display -pre.1 as "unanticipated"!

So I think prerelease handling is as good as it was before now, and
possibly better in cases we haven't thought of, and I get to remove the
special options to isConstraintSatisfied!

Also kill the old tests that called into the old Resolver directly.
The prerelease test was the only meaty one, and it's ported now.