Delete mongod.lock file when starting Meteor
When we try to kill old running versions of mongod, then we may have to
resort to killing the process without it having an opportunity to close
gracefully, so we need to also delete the lock file if we want to be able
to start mongod again.

This is made especially worse on Windows, where `process.kill` always
kills a process non-gracefully.

Should fix https://github.com/meteor/windows-preview/issues/34, and some
self-test failures.