Improve Blaze performance on initial array rendering
This is one of a longer sequence of changes we can make to make
the diffing algorithm in Blaze be linear as opposed to quadratic
which it is at the momeny, in general.

This change specifically makes it so that the very common case of
an {{#each}} over an array that switches from empty to a long array
much faster, since the adds in those case are always at the end
of the array. This means there's no bookkeeping necessary to
update indexes.

This changes reduces the rendering time from the example posted
on #3596 from 5000ms to 1729ms.