Fix #2534 for oplog too
In order to test this, we applied the following diff to
minimongo/selector.js:

--- a/packages/minimongo/selector.js
+++ b/packages/minimongo/selector.js
@@ -591,6 +591,8 @@ ELEMENT_OPERATORS = {
   },
   $in: {
     compileElementSelector: function (operand) {
+      if (operand === null)
+        operand = [];
       if (!isArray(operand))
         throw Error("$in needs an array");

This means that minimongo now allows `$in: null`, and thus
OplogObserveDriver can be used for the query in the new 'bad query'
test, but mongod still returns an error that is processed by the passing
test.  (The client side of the test fails with this temporary change,
because minimongo no longer throws!)

It's hard to see how to test this better, because minimongo should throw
on any queries that mongod throws on.