minor refactor of the way we handle update/insert/remove requests
Instead of one loop that iterates over each of the three types of
requests and then performs different operations, separate into a much
more reasonable construct of three different functions that share some
calls to internal functions.

Added an ability to check that an input is a function in match.
Unit tests will be added in a separate commit.