bulk inserts
Propagate bulk insert ability into the mongo driver, as well as inserts and validated
inserts in collection.js.

We are intentionally keeping the APIs for mongo_driver backwards compatible, since that
function gets called from multiple places in the code and is expected to mirror
Mongo's actual API (which takes both arrays and individual documents).