separating another set of update-remove-insert loops into separate methods
The idea of using a loop to define update/insert/remove methods is kind of cool,
and separating them out does result in a little bit of repeated code. But it is
much, much easier to interact with this code when it is separated out, arguments are named,
etc. Plus, moving some stuff out of the function to make the code somewhat more
clear (IMO).