Fix glasser's "actual bug"
That's what we've been calling it.

To see this break and then work, add the following line above
the comment that begins `// lock down versions...`:

```
self.solution = logic.solveAssuming('bar');
```

And then run the test "previous solution no longer needed"
in input-tests.js.

The old code would lock down "bar", including the fact that it
is in the current solution.  The new code only locks down the version
number in the case where "bar" is used, so it is able to be optimized
away.