`meteor update` with no args patches indirect deps
With this change, `meteor update` with no arguments, in addition to
updating direct dependencies, updates patch versions of indirect
dependencies, so you will get your version 1.4.2 in the above example.

Background: As of the new solver, we already prefer patched versions 
when selecting a version for a new indirect dependency.  For example,
if 1.4.0 is the oldest version we could choose, we will take 1.4.1
instead if available.  However, if 1.4.2 came out, there would never
be an occasion to take it, unless you explicitly typed
`meteor update the-indirect-dependency`.  `meteor update` with no
args expands to `meteor update <all direct deps>`.

Also improve comments, tests, and code around CS.Input.