Let `meteor update --release` downgrade packages
When the tool runs the constraint solver to see what releases can
be chosen, it creates its own ProjectContext and does not pass
command-line flags through.  Additionally, if the release you ask
for is not possible to choose for whatever reason (constraints in
.meteor/packages, and now the guard against downgrading root
dependencies), the error is swallowed and a pleasant "success"
message is printed.  We should probably fix this at some point.

For now, `meteor update --release` just runs the constraint solver
as if `--allow-incompatible-update` was passed, which effectively
means it gets to ignore `.meteor/versions` if necessary.