Add a self-test for tarballs with long paths.
This is important for Windows, where there's a 260 character on path length.
Windows has a "cheat code" where directories prefixed by \\?\ are allowed
to have length of up to 64K. This self-test will help us make sure all of
the libraries we use in Windows (for example fstream) correctly use \\?\
as necessary to overcome the limitation.